binary, dot2, dot3
