+++ src/ruby/nodes/conditionals.ts
@@ -193,7 +193,7 @@ function canTernaryStmts(stmts: Ruby.Stmts) {
 
   // If the user is using one of the lower precedence "and" or "or" operators,
   // then we can't use a ternary expression as it would break the flow control.
-  if (stmt.type === "binary" && ["and", "or"].includes(stmt.body[1])) {
+  if (stmt.type === "binary" && ["and", "or"].includes(stmt.operator)) {
     return false;
   }
