+++ src/utils/containsAssignment.ts
@@ -1,8 +1,9 @@
 import type { Ruby } from "../types";
+import getChildNodes from "./getChildNodes";
 
 // If the node is a type of assignment or if the node is a paren and nested
 // inside that paren is a node that is a type of assignment.
-function containsAssignment(node: Ruby.AnyNode | Ruby.Stmts) {
+function containsAssignment(node: null | Ruby.AnyNode | Ruby.Stmts) {
   if (!node) {
     return false;
   }
@@ -11,8 +12,7 @@ function containsAssignment(node: Ruby.AnyNode | Ruby.Stmts) {
     return true;
   }
 
-  const anyNode = node as any;
-  return Array.isArray(anyNode.body) && anyNode.body.some(containsAssignment);
+  return getChildNodes(node).some(containsAssignment);
 }
 
 export default containsAssignment;
